Login fixed. User homepage frontend.

+++ src/main/java/it/polimi/dbproject/entities/EmployeeEntity.java
@@ -15,7 +15,7 @@ public class EmployeeEntity implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "employee_id", nullable = false)
-    private Long employeeId;
+    private int employeeId;
 
     @Column(name = "email", unique = true, nullable = false)
     private String email;
@@ -38,11 +38,11 @@ public class EmployeeEntity implements Serializable {
 
     }
 
-    public Long getEmployee_id() {
+    public int getEmployee_id() {
         return employeeId;
     }
 
-    public void setEmployee_id(Long employeeId) {
+    public void setEmployee_id(int employeeId) {
         this.employeeId = employeeId;
     }
